Backboner API

Backboner.Backbone — Type

Backbone{T<:Real,M<:AbstractMatrix{T}} <: AbstractVector{AbstractVector{T}}

The Backbone type is designed to efficiently store and manipulate the three-dimensional coordinates of backbone points.

Backboner.ChainedBonds — Type

ChainedBonds{T<:Real,V<:AbstractVector{T}}

A lazy way to store a backbone as a series of bond lengths, bond angles, and torsion_angles.

Examples

julia> backbone = Backbone(randn(Float32, 3, 660));

julia> bonds = ChainedBonds(backbone)
ChainedBonds{Float64, Vector{Float64}} with 659 bond lengths, 658 bond angles, and 657 torsion angles

Backboner.append_bonds — Method

append_bonds(backbone, bond_lengths, bond_angles, torsion_angles)

Backboner.get_displacements — Method

get_displacements(backbone, start, step, stride)

Get the displacements between points in backbone, where start is the index of the first point, step is the offset of the other points along the backbone (e.g. 1 for bond vectors), and stride is the number of points to skip after each step.

Backboner.get_distances — Method

get_distances(backbone, start, step, stride)

Get the Euclidean distances between points in backbone, where start is the index of the first point, step is the offset of the other points along the backbone (e.g. 1 for bond lengths), and stride is the number of points to skip after each step.

Backboner.is_knotted — Function

is_knotted(backbone::Backbone)

Check if a backbone is knotted.

Backboner.prepend_bonds — Method

prepend_bonds(backbone, bond_lengths, bond_angles, torsion_angles)