Backboner API

Backboner.Backbone — Type

Backbone{T <: Real} <: AbstractMatrix{T}

The Backbone type is designed to efficiently store and manipulate the three-dimensional coordinates of backbone atoms.

Examples

A Backbone can be created from a matrix of coordinates:

julia> backbone = Backbone(zeros(3, 5)) # 5 atoms with 3 coordinates each
3×5 Backbone{Float64, Matrix{Float64}}:
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

julia> backbone[1] = [1.0, 2.0, 3.0]; # set the first atom's coordinates

julia> backbone
3×5 Backbone{Float64, Matrix{Float64}}:
 1.0  0.0  0.0  0.0  0.0
 2.0  0.0  0.0  0.0  0.0
 3.0  0.0  0.0  0.0  0.0

julia> backbone[1:2] # indexing by range returns a new Backbone
3×2 Backbone{Float64, Matrix{Float64}}:
 1.0  0.0
 2.0  0.0
 3.0  0.0

Arrays will always be flattened to a 3xN matrix:

julia> backbone = Backbone(zeros(3, 3, 4)) # e.g. 3 coordinates per atom, 3 atoms per residue, 4 residues
3×12 Backbone{Float64, Matrix{Float64}}:
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

Backboner.ChainedBonds — Type

ChainedBonds{T <: Real, V <: AbstractVector{T}}

A lazy way to store a backbone as a series of bond lengths, angles, and dihedrals. It can be instantiated from a Backbone or a matrix of bond vectors. It can also be used to instantiate a Backbone using the Backbone(bonds::ChainedBonds) constructor.

Examples

julia> backbone = Protein.readpdb("test/data/1ZAK.pdb")["A"].backbone
3×660 Backbone{Float32, Matrix{Float32}}:
 22.346  22.901  23.227  24.115  24.478  25.289  26.091  26.814  …  23.137  22.572  21.48   22.041  21.808  22.263  21.085
 17.547  18.031  16.793  16.923  15.779  14.65   14.958  13.827     13.041  14.235  14.668  14.866  13.861  13.862  14.233
 23.294  21.993  21.163  20.175  19.336  20.009  21.056  21.652      5.676   5.844   4.974   3.569   2.734   1.355   0.446

julia> bonds = ChainedBonds(backbone)
ChainedBonds{Float32, Vector{Float32}} with 659 bonds, 658 angles, and 657 dihedrals

Backboner.Frame — Type

Frame{T <: Real}

A Frame is a combination of a rotation and a translation, which can be applied to a set of coordinates.

Backboner.Frames — Type

Frames{T <: Real, M <: AbstractMatrix{T}} <: AbstractVector{Frame{T}}

The Frames type is designed to efficiently store and manipulate the rotation and translation of a set of Frames.

Backboner.append_bonds — Method

append_bonds(backbone, lengths, angles, dihedrals)

Backboner.idealize — Function

Note

Zygote must be imported in order to activate the ZygoteIdealizationExt extension, which defines the idealize(::Backbone) method.

Backboner.is_knotted — Function

is_knotted(backbone::Backbone)

Check if a backbone is knotted.